# contracts/openapi.yaml
openapi: 3.1.1
info:
  title: Tawsel public contract — identity, intake, planning and execution
  version: 0.1.0
  description: P07–P21 locally implemented identity, intake, locations, planning,
    round start, current activity, outcomes, eligibility, workday closure,
    device takeover and source-branch return receipt/disposition. Provider,
    browser and PostgreSQL evidence are distinguished in phase ledgers. Branch
    resume/redispatch, corrections, full offline capture and signed delivery
    remain designed.
  license:
    name: Private development contract; no public release
...
paths:
...
  /api/v1/current/arrival:
    post:
      operationId: current.recordArrival
      summary: current.recordArrival
      tags:
        - Current activity
      security:
        - companySession: []
        - personalSession: []
      parameters:
        - name: kind
          in: query
          required: true
          schema:
            type: string
            enum:
              - personal
              - company
        - name: X-CSRF-Token
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CurrentArrivalCommand"
      responses:
        "200":
          description: Authoritative response; pending is not acceptance.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentActionResult"
        "400":
          description: Invalid request
        "401":
          description: Session required
        "403":
          description: Current scope or CSRF denied
        "409":
          description: Retained activity, ownership, eligibility or relevant revision
            rejection. Query or retry the same stable action ID.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentActionResult"
...
  /api/v1/returns/request:
    post:
      operationId: return.requestHandover
      tags:
        - Returns
      summary: Driver offers source-branch pieces; request is not receipt/stock.
      security:
        - companySession: []
        - personalSession: []
      parameters:
        - name: kind
          in: query
          required: true
          schema:
            type: string
            enum:
              - company
              - personal
          description: Company only; personal is rejected by domain rules.
        - name: X-CSRF-Token
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReturnRequestCommand"
      responses:
        "200":
          description: Committed result or authoritative scoped read.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReturnActionResult"
        "400":
          description: Invalid canonical input.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "401":
          description: Authentication required.
        "403":
          description: Scope/capability/CSRF denied.
        "404":
          description: Resource unavailable.
        "409":
          description: Stale item, wrong source branch, incompatible lifecycle or excess
            quantity. Held goods are unchanged.
        "503":
          description: No assumed receipt. Recover the action and retry the same envelope.
  /api/v1/returns/groups:
    get:
      operationId: return.listSourceBranchGroups
      tags:
        - Returns
      summary: Group held return-required portions by originating branch.
      security:
        - companySession: []
        - personalSession: []
      parameters:
        - name: kind
          in: query
          required: true
          schema:
            type: string
            enum:
              - company
              - personal
          description: Company only; personal is rejected by domain rules.
      responses:
        "200":
          description: Committed result or authoritative scoped read.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReturnGroups"
        "400":
          description: Invalid canonical input.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "401":
          description: Authentication required.
        "403":
          description: Scope/capability/CSRF denied.
        "404":
          description: Resource unavailable.
        "409":
          description: Stale item, wrong source branch, incompatible lifecycle or excess
            quantity. Held goods are unchanged.
        "503":
          description: No assumed receipt. Recover the action and retry the same envelope.
  /api/v1/returns/requests/{requestId}:
    get:
      operationId: return.getRequest
      tags:
        - Returns
      summary: Scoped request, offered/received/unresolved subsets and revisions.
      security:
        - companySession: []
        - personalSession: []
      parameters:
        - name: kind
          in: query
          required: true
          schema:
            type: string
            enum:
              - company
              - personal
          description: Company only; personal is rejected by domain rules.
        - name: requestId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Committed result or authoritative scoped read.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReturnRequestView"
        "400":
          description: Invalid canonical input.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "401":
          description: Authentication required.
        "403":
          description: Scope/capability/CSRF denied.
        "404":
          description: Resource unavailable.
        "409":
          description: Stale item, wrong source branch, incompatible lifecycle or excess
            quantity. Held goods are unchanged.
        "503":
          description: No assumed receipt. Recover the action and retry the same envelope.
  /api/v1/returns/actions/{actionId}:
    get:
      operationId: return.getResult
      tags:
        - Returns
      summary: Recover own driver offer command; pending is not physical receipt.
      security:
        - companySession: []
        - personalSession: []
      parameters:
        - name: kind
          in: query
          required: true
          schema:
            type: string
            enum:
              - company
              - personal
          description: Company only; personal is rejected by domain rules.
        - name: actionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Committed result or authoritative scoped read.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReturnActionStatus"
        "202":
          description: No committed action result visible.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReturnActionStatus"
        "400":
          description: Invalid canonical input.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "401":
          description: Authentication required.
        "403":
          description: Scope/capability/CSRF denied.
        "404":
          description: Resource unavailable.
        "409":
          description: Stale item, wrong source branch, incompatible lifecycle or excess
            quantity. Held goods are unchanged.
        "503":
          description: No assumed receipt. Recover the action and retry the same envelope.
  /api/v1/returns/requests/{requestId}/confirmation:
    post:
      operationId: return.checkConfirmation
      tags:
        - Returns
      summary: Read server confirmation for an explicit cumulative claimed subset.
        Waiting never grants resume; P22 rechecks under lock.
      security:
        - companySession: []
        - personalSession: []
      parameters:
        - name: kind
          in: query
          required: true
          schema:
            type: string
            enum:
              - company
              - personal
          description: Company only; personal is rejected by domain rules.
        - name: requestId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-CSRF-Token
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReturnConfirmationQuery"
      responses:
        "200":
          description: Committed result or authoritative scoped read.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReturnConfirmation"
        "400":
          description: Invalid canonical input.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "401":
          description: Authentication required.
        "403":
          description: Scope/capability/CSRF denied.
        "404":
          description: Resource unavailable.
        "409":
          description: Stale item, wrong source branch, incompatible lifecycle or excess
            quantity. Held goods are unchanged.
        "503":
          description: No assumed receipt. Recover the action and retry the same envelope.
  /api/v1/erp/returns/commands/return.confirmSubsetReceipt:
    post:
      operationId: return.confirmSubsetReceipt
      tags:
        - Returns
      summary: Native ERP trusted actor confirms actually received requested subset.
      security:
        - ProvisioningService: []
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReturnReceiveCommand"
      responses:
        "200":
          description: Committed result or authoritative scoped read.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReturnActionResult"
        "400":
          description: Invalid canonical input.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "401":
          description: Authentication required.
        "403":
          description: Scope/capability/CSRF denied.
        "404":
          description: Resource unavailable.
        "409":
          description: Stale item, wrong source branch, incompatible lifecycle or excess
            quantity. Held goods are unchanged.
        "503":
          description: No assumed receipt. Recover the action and retry the same envelope.
  /api/v1/erp/returns/commands/return.recordDisposition:
    post:
      operationId: return.recordDisposition
      tags:
        - Returns
      summary: ERP loss/damage disposition distinct from physical receipt and inventory.
      security:
        - ProvisioningService: []
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReturnDisposeCommand"
      responses:
        "200":
          description: Committed result or authoritative scoped read.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReturnActionResult"
        "400":
          description: Invalid canonical input.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "401":
          description: Authentication required.
        "403":
          description: Scope/capability/CSRF denied.
        "404":
          description: Resource unavailable.
        "409":
          description: Stale item, wrong source branch, incompatible lifecycle or excess
            quantity. Held goods are unchanged.
        "503":
          description: No assumed receipt. Recover the action and retry the same envelope.
  /api/v1/erp/returns/pending:
    get:
      operationId: return.listPending
      tags:
        - Returns
      summary: Native source-scoped pending requests for one driver and originating
        branch, paginated in pages of 100 without a returns quota.
      security:
        - ProvisioningService: []
      parameters:
        - name: driverId
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: sourceBranchId
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: cursor
          in: query
          required: false
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Committed result or authoritative scoped read.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReturnRequestList"
        "400":
          description: Invalid canonical input.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "401":
          description: Authentication required.
        "403":
          description: Scope/capability/CSRF denied.
        "404":
          description: Resource unavailable.
        "409":
          description: Stale item, wrong source branch, incompatible lifecycle or excess
            quantity. Held goods are unchanged.
        "503":
          description: No assumed receipt. Recover the action and retry the same envelope.
  /api/v1/erp/returns/requests/{requestId}:
    get:
      operationId: return.getNativeRequest
      tags:
        - Returns
      summary: Native source-scoped accurate
        requested/received/unresolved/lost/damaged state per item.
      security:
        - ProvisioningService: []
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Committed result or authoritative scoped read.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReturnRequestView"
        "400":
          description: Invalid canonical input.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "401":
          description: Authentication required.
        "403":
          description: Scope/capability/CSRF denied.
        "404":
          description: Resource unavailable.
        "409":
          description: Stale item, wrong source branch, incompatible lifecycle or excess
            quantity. Held goods are unchanged.
        "503":
          description: No assumed receipt. Recover the action and retry the same envelope.
  /api/v1/erp/returns/actions/{actionId}:
    get:
      operationId: return.getNativeResult
      tags:
        - Returns
      summary: Recover durable native source receipt/disposition command results; 202
        means unknown pending.
      security:
        - ProvisioningService: []
      parameters:
        - name: actionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Committed result or authoritative scoped read.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReturnActionStatus"
        "202":
          description: No committed action result visible.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReturnActionStatus"
        "400":
          description: Invalid canonical input.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "401":
          description: Authentication required.
        "403":
          description: Scope/capability/CSRF denied.
        "404":
          description: Resource unavailable.
        "409":
          description: Stale item, wrong source branch, incompatible lifecycle or excess
            quantity. Held goods are unchanged.
        "503":
          description: No assumed receipt. Recover the action and retry the same envelope.
  /api/v1/branches/commands/branch.interruptRound:
    post:
      operationId: branch.interruptRound
      tags:
        - Returns
      summary: Pause heading sequence and enter source-branch service within same
        round/capacity.
      security:
        - companySession: []
        - personalSession: []
      parameters:
        - name: kind
          in: query
          required: true
          schema:
            type: string
            enum:
              - company
              - personal
          description: Company only; personal is rejected by domain rules.
        - name: X-CSRF-Token
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BranchInterruptCommand"
      responses:
        "200":
          description: Committed result or authoritative scoped read.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"
        "400":
          description: Invalid canonical input.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "401":
          description: Authentication required.
        "403":
          description: Scope/capability/CSRF denied.
        "404":
          description: Resource unavailable.
        "409":
          description: Stale item, wrong source branch, incompatible lifecycle or excess
            quantity. Held goods are unchanged.
        "503":
          description: No assumed receipt. Recover the action and retry the same envelope.
  /api/v1/branches/commands/branch.recordArrival:
    post:
      operationId: branch.recordArrival
      tags:
        - Returns
      summary: Explicit arrival at the source-bound branch; records physical origin
        without inferring receipt.
      security:
        - companySession: []
        - personalSession: []
      parameters:
        - name: kind
          in: query
          required: true
          schema:
            type: string
            enum:
              - company
              - personal
          description: Company only; personal is rejected by domain rules.
        - name: X-CSRF-Token
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BranchArrivalCommand"
      responses:
        "200":
          description: Committed result or authoritative scoped read.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"
        "400":
          description: Invalid canonical input.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "401":
          description: Authentication required.
        "403":
          description: Scope/capability/CSRF denied.
        "404":
          description: Resource unavailable.
        "409":
          description: Stale item, wrong source branch, incompatible lifecycle or excess
            quantity. Held goods are unchanged.
        "503":
          description: No assumed receipt. Recover the action and retry the same envelope.
  /api/v1/branches/commands/branch.resumeRound:
    post:
      operationId: branch.resumeRound
      tags:
        - Returns
      summary: Resume retained sequence after authoritative claimed-subset receipt; no
        whole-batch gate.
      security:
        - companySession: []
        - personalSession: []
      parameters:
        - name: kind
          in: query
          required: true
          schema:
            type: string
            enum:
              - company
              - personal
          description: Company only; personal is rejected by domain rules.
        - name: X-CSRF-Token
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BranchResumeCommand"
      responses:
        "200":
          description: Committed result or authoritative scoped read.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"
        "400":
          description: Invalid canonical input.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "401":
          description: Authentication required.
        "403":
          description: Scope/capability/CSRF denied.
        "404":
          description: Resource unavailable.
        "409":
          description: Stale item, wrong source branch, incompatible lifecycle or excess
            quantity. Held goods are unchanged.
        "503":
          description: No assumed receipt. Recover the action and retry the same envelope.
...
components:
  schemas:
...
    Problem:
      $ref: ./common.schema.json#/$defs/Problem
...
    ActionResult:
      $ref: ./action-result.v1.schema.json
...
    CurrentArrivalCommand:
      $ref: ./current-activity.schema.json#/$defs/ArrivalCommand
...
    CurrentActionResult:
      $ref: ./current-activity.schema.json#/$defs/ActionResult
...
    ReturnRequestCommand:
      $ref: ./returns.schema.json#/$defs/RequestCommand
    ReturnReceiveCommand:
      $ref: ./returns.schema.json#/$defs/ReceiveCommand
    ReturnDisposeCommand:
      $ref: ./returns.schema.json#/$defs/DisposeCommand
...
    ReturnRequestView:
      $ref: ./returns.schema.json#/$defs/RequestView
    ReturnRequestList:
      $ref: ./returns.schema.json#/$defs/RequestList
...
    ReturnGroups:
      $ref: ./returns.schema.json#/$defs/Groups
...
    ReturnConfirmationQuery:
      $ref: ./returns.schema.json#/$defs/ConfirmationQuery
    ReturnConfirmation:
      $ref: ./returns.schema.json#/$defs/Confirmation
...
    ReturnActionResult:
      $ref: ./returns.schema.json#/$defs/ActionResult
    ReturnActionStatus:
      $ref: ./returns.schema.json#/$defs/ActionStatus
...
    BranchInterruptCommand:
      $ref: ./branch-activity.schema.json#/$defs/InterruptCommand
    BranchArrivalCommand:
      $ref: ./branch-activity.schema.json#/$defs/ArrivalCommand
    BranchResumeCommand:
      $ref: ./branch-activity.schema.json#/$defs/ResumeCommand
...
  securitySchemes:
    companySession:
      type: apiKey
      in: cookie
      name: __Host-tawsel-company
    personalSession:
      type: apiKey
      in: cookie
      name: __Host-tawsel-personal
    ProvisioningService:
      type: http
      scheme: bearer
      description: twp_<credential UUID>.<32 random bytes lowercase hex>. Hashed
        server-side; tenant/integration bound, expiring, rotatable. TLS required
        outside loopback.
    ProvisioningOperator:
      type: http
      scheme: bearer
      description: Out-of-band server operator token; bootstrap and recovery rotation
        only. Never distribute to ERP application users.
...